requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python
pydantic-settings
paddleocr[all]
ollama
onnxruntime
piper-phonemize
